added system layout container
