+++ package.json
@@ -29,7 +29,7 @@
     "react-intl": "^5.4.3",
     "react-tooltip": "^4.5.1",
     "sharp": "^0.31.2",
-    "swr": "^1.3.0",
+    "swr": "^2.0.4",
     "usehooks-ts": "^2.9.1",
     "yup": "^0.32.11"
   },
