--- src/modules/systems/catalogueItemSection/types/catalogueItemSection.ts
@@ -1,12 +1,0 @@
-import { FieldValues } from 'react-hook-form'
-
-export interface SystemItemFormType extends FieldValues {
-  catalogueItemUID?: string
-  itemUsageUID: string
-  eun: string
-  name: string
-  serialNumber: string
-  batchNumber: string
-  obsolete: string
-  estimatedLifeTimeMonths: string
-}
