+++ src/modules/systems/components/catalogueItemSection/CatalogueItemSection.cont.tsx
