+++ src/modules/systems/components/catalogueItemSection/componenets/CatalogueItemModal.tsx
